Convert hosting device's created_at attribute to datetime object only if needed
Initially the hosting device['created_at'] atribute is coming as a string. Since we use it all the time, it is converted to a datetime object. But when the router_deleted action is happening, we inadvertently converts it again causing the TypeError. An instanceof() check is added to
avoid this.

Closes defect: DE1016

Change-Id: Iadd1ae4eb58a75452d958094a185d989f7ac9a1b