Retry if HA group info is not found for HA enabled router
This patch changes how HA information is added to ports for
an HA enabled router. If the HA group for a router and a particular
subnet is not found in the DB, the plugin will retry the DB
query a few times in hope that the HA group DB entry will be
commited to the DB during those retries.

The patchs also makes a few minor code cleanups of the code
handling the HA group info, in particular make use of a constant
defined in the HA extension instead of its string value directly.

Partially closes bug: DE936

Change-Id: Ic5ff8b1b3d4ba97a5aec995930f47e775c3a030c