DE971 - Update to handle initial unsuccessful sync_devices
In this patch, a "sync_devices" retry is introduced in order to
handle the case where the agent notifies the plugin that a hosting
device has been revived, but the plugin hasn't rescheduled
any routers yet.  On the agent side, the device_sync may not happen
because the routers fetched from the plugin for a particular device
might be blank/empty.  In this case, the fix keeps the sync_devices
buffer unmodified in the routing_svc_helper which allows subsequent
process_service task attempts to retry.