DE1243 - Corrected key exception
This fix corrects an exception raised when the cfg-syncer encounters
an existing running-cfg for the ext-gw interface.

In this specific case, neutron-db is empty (aside from the global routers)
being present.  The criteria to have the segment_nat_dict populated is not
met if there isn't any tenant router configured with both a private net
and ext-net which then leads to the key lookup error when performing
clean_interfaces.