Changes logging of transient issue from error to debug type
The Cisco L3 router plugin regularly need to fetch hosting
information for routers. That information may transiently
not be available because it has not yet been persisted to
the DB. The DB lookup then fails in which case this is
logged as an error.

The L3 router plugin does however handle this gracefully
and will make renewed attempts later, which will eventually
succeed. It is therefore better to log as debug rather than
error as the latter may lead the operator to think there
is a real error that needs to be acted on.

Fixes defect: DE1768

Change-Id: Ie085179e7f28e97c809cc895f6c83a222dbbe6b3